compiler: Drop dead code
