+++ devito/ir/support/__init__.py
@@ -1,7 +1,6 @@
 from .vector import *  # noqa
 from .basic import *  # noqa
 from .space import *  # noqa
-from .stencil import *  # noqa
 from .guards import *  # noqa
 from .utils import *  # noqa
 from .properties import *  # noqa
